Renamed MedicalSummary to MedicationSummary, moved MedicationItem into MedicationSummary as an static class, updated FLTX medicalSummary part for simpler conditions in If-statements. Added tests for PS-mapper.
